text_adventure/requirements.txt
langchain-community==0.0.19
langchain-core==0.1.21
langchain-openai==0.0.5
langchain==0.1.5
langsmith==0.0.87
python-dotenv==1.0.1
openai==1.11.1
streamlit==1.29.0
cassandra-driver==3.29.1
cassio==0.1.5
astrapy==1.0.0
astroid==2.15.5
asttokens==2.2.1
astunparse==1.6.3
requests==2.31.0
requests-oauthlib==1.3.1
